portfolio/requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
flask-cors==6.0.2